+++ upkie/utils/__init__.py
@@ -3,3 +3,6 @@
 #
 # SPDX-License-Identifier: Apache-2.0
 # Copyright 2023 Inria
+#
+## \namespace upkie.utils
+## \brief Utility functions.