+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- envs: Register UpkieServoPositions and UpkieServoTorques environments
+
 ## [5.1.0] - 2024-08-14
 
 ### Added