+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- BulletInterface: Fix application of external forces
 - envs: Register UpkieServoPositions and UpkieServoTorques environments
 
 ## [5.1.0] - 2024-08-14