+++ CHANGELOG.md
@@ -9,7 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Added
 
-- BulletInterface: add torque noise to joint-property configuration
+- BulletInterface: add torque control noise to joint-property configuration
+- BulletInterface: add torque measurement noise to joint-property configuration
 - examples: Apply an external force to lift an Upkie in sim
 - examples: Simulation with joint friction and noise