+++ CHANGELOG.md
@@ -24,6 +24,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - CICD: Enable stable-baselines3 environment-check unit tests
 - docs: Add missing documentation pages
 - docs: Document Gym environment wrappers
+- envs: Handle spine errors raised in base env constructor
 - envs: Register UpkieServoPositions and UpkieServoTorques environments
 
 ### Removed