+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [5.2.0] - 2024-09-30
+
 ### Added
 
 - BulletInterface: Apply custom initial joint configuration, if provided
@@ -628,7 +630,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 Starting this changelog.
 
-[unreleased]: https://github.com/upkie/upkie/compare/v5.1.0...HEAD
+[unreleased]: https://github.com/upkie/upkie/compare/v5.2.0...HEAD
+[5.2.0]: https://github.com/upkie/upkie/compare/v5.1.0...v5.2.0
 [5.1.0]: https://github.com/upkie/upkie/compare/v5.0.1...v5.1.0
 [5.0.1]: https://github.com/upkie/upkie/compare/v5.0.0...v5.0.1
 [5.0.0]: https://github.com/upkie/upkie/compare/v4.0.0...v5.0.0